Robustize the processing of newly occurred problems
Problem Service announces changes in a problem in same way as occurrence
of a new problem. The problems cache wasn't designed according to this
approach and with each change of already processed problem, the cache
produced a warning about attempt to push a known problem to the cache.

This change makes the cache permissive to pushing known problems to the
cache. The cache still detects this case but forces the problem to
reload itself instead of producing the warning.

Signed-off-by: Jakub Filak <jfilak@redhat.com>