Move gnome_abrt module check to module's Makefile
This commit fixes problems with pylint 0.26. For some unknown reason
pylint reenabled a locally disabled warning R0923 in gnome-abrt script
if the executable and the module were passed both in command line
arguments. Maybe it is intentional behaviour, maybe not. I do not
care, this patch makes the behaviour understandable to me.

Signed-off-by: Jakub Filak <jfilak@redhat.com>