Load only the most recent reported to value
If reported_to contains more "Bugzilla: URL" entries, load only the last
one. This patch is supposed to avoid showing multiple entries for "ABRT
Server: URL". It is expected that reporter-ureport always adds "ABRT
Server: URL" to reported_to file (useful when the crash is unknown, it
will be the only one URL in the reported to file) but FAF
returns the similar url itself if the crash is known.

Related to rhbz#977029

Signed-off-by: Jakub Filak <jfilak@redhat.com>