Use human readable type string everywhere
Richard Marko <rmarko@redhat.com> came up with a great idea of using
human readable type strings instead of confusing machine strings.

He has implemented his idea in the patch adding conversion of 'CCpp' to
'C/C++' in the list of problems. But there is another place where we
want to display this kind of string.

This commit extends Richard's commit[1] a bit.

1: 237654ca7f1d6fbb95690e9f8c56b2d6891d4f96

Signed-off-by: Jakub Filak <jfilak@redhat.com>