Do not pass None to function expecting str object
The C wrapper requires a str object. We can fix this by checking
validity of the argument in the function or we can just say that you
must pass a valid str object. I have chosen the former variant
because adding the check to the C wrapper wouldn't be so trivial
(for me).

"cmdline" is True if it is not None and is not empty. There is no need
to call the function with an empty string and the condition is easier
to comprehend.

Fixes #152

Signed-off-by: Jakub Filak <jfilak@redhat.com>