Remove Expert mode and the remaining Analyze code
The Expert mode was originally introduced together with the Analyze
feature by the commit 532cbd4. The Expert mode was intended to hide
the Analyze feature by default and show in only when the Expert mode
is on. Later the commit f5eb93f removed most of the Analyze feature
although some of its code still remained and was unused. At the same
time, probably by mistake, the Expert mode was switched to control
the visibility of the Details context menu item.

This commit removes the Expert mode which is unused and implemented
incorrectly and any remains of the Analyze feature. Closes #182.