views: Fix side pane sizing
Currently, the code adds the double of the X offset of box_header_left
to determine the minimum width of box_panel_left. That results in brokenness
when GtkSettings:gtk-shell-shows-app-menu is set to false or the theme adds
actual padding or margins. That is fixable by getting the padding and
margins from the style context and adding them up.

Closes https://github.com/abrt/gnome-abrt/issues/213

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>