views: Prevent ellipsization of app label
ProblemRow contains a GtkGrid, which is mistakenly set to have
homogeneous columns in order to have proper alignment of the children.
However, this results in the application label being ellipsized if it’s
long enough. To fix that, this commit unsets column-homogeneous and sets
the application and type labels to be expanded horizontally.

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>