views: Set up actions earlier
Currently, actions are added to the window *after* populating the
problem list, which is when insensitivization of the report button
happens (if the problem is unreportable). However, since the report
button is given an action name, the property will be flipped again if
the action is enabled (which it is).

Another solution is to also conditionally disable the actions.

Fixes https://github.com/abrt/gnome-abrt/issues/242

Signed-off-by: Ernestas Kulik <ekulik@redhat.com>