wrappers: Avoid multiple-definition of _PYGObject_API
Given that GCC 10 defaults to -fno-common and that such code violates
the C spec, we need to add a workaround, since PyGObject works in
mysterious ways.

We define NO_IMPORT_PYGOBJECT where pygobject.h is included, so that
the struct is declared with external linkage, but not in module,c, since
we need at least one symbol defined.

https://bugzilla.redhat.com/show_bug.cgi?id=1793523