Fix up conditional linting
We were looking for the pylint binary while we only need the Python
module. Moreover, the spec file wasn’t actually making use of the
option. This commit changes the type of the option from feature to
boolean, since we cannot easily detect the existence of the module and
renames the option to be a verb, because it makes more sense this way
now.