Fix Python module installation
Currently, all sources are installed as “pure”, which does not apply to
the wrapper module. Moreover, Meson previously defaulted to “false” for
the “pure” kwarg, which was subsequently fixed. Both of these issues
make the RPM package unbuildable, which can be fixed by making all
sources unpure (they all share the same module name, so it’s either this
or renaming the wrapper module to avoid import conflicts).

https://github.com/mesonbuild/meson/commit/efb86088bcf89