views: Unconditionally use Cairo surface for application icon
The current code is all sorts of wrong: it uses explicit None returns
when no caller uses the return value, it needlessly avoids using Cairo
surfaces for icon data and it performs some pointless checks on
arguments. This commit refactors the helper function and moves it out to
the only place where it’s used.