views: Replace self-rolled function for human-readable dates
The list view uses a different date humanization method than what is
shown in the oops-window as "First detected". Reduce complexity and
improve the label in the list view by using the same functionality for
both.

After the turn of the year, problems that occurred a few days ago were
listed as "Last year", which is technically correct, but not very
helpful.